requirements.txt
fastapi
uvicorn
requests
python-dotenv

openai

duckduckgo_search
trafilatura
